% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barley_area}
\alias{barley_area}
\title{Barley leaf area example data}
\format{
A list with three elements:
\describe{
  \item{gwas}{GWAS results (data frame) for leaf area}
  \item{hapmap}{Genotypes as returned by [readHapmap()] (list of data frames per chromosome)}
  \item{pheno}{Phenotypes: accession name (`Taxa`) and leaf area (`Area`, cm2)}
}
}
\source{
ICARDA barley breeding program.
}
\usage{
barley_area
}
\description{
GWAS results, genotypes and phenotypes for leaf area in a barley panel
(275 accessions, Barley 50K SNP array). The same data are available as
files in `system.file("extdata", package = "HaploTraitR")`.
}
\keyword{datasets}
