% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_geno_pheno.R
\name{boxplot_geno_pheno}
\alias{boxplot_geno_pheno}
\title{Generate boxplots for genotype-phenotype data}
\usage{
boxplot_geno_pheno(geno_pheno, outfolder = NULL)
}
\arguments{
\item{geno_pheno}{A data frame containing genotype-phenotype data}

\item{outfolder}{An optional output folder for saving the plots}
}
\value{
A list of ggplot2 objects or NULL if outfolder is specified
}
\description{
Generate boxplots for genotype-phenotype data
}
