% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_haplotypes.R
\name{boxplot_genotype_phenotype}
\alias{boxplot_genotype_phenotype}
\title{Generate boxplots of the trait for each significant GWAS SNP}
\usage{
boxplot_genotype_phenotype(
  pheno,
  gwas,
  hapmap,
  plot_width = 5,
  plot_height = 5.5,
  plot_all = TRUE
)
}
\arguments{
\item{pheno}{A data frame containing phenotype data (accession, trait)}

\item{gwas}{A list of significant GWAS SNPs (from [filter_gwas_data()])}

\item{hapmap}{A list of HapMap data frames (from [readHapmap()])}

\item{plot_width}{The width of the saved plots (inches)}

\item{plot_height}{The height of the saved plots (inches)}

\item{plot_all}{If `TRUE` plot all SNPs, otherwise only SNPs with a
significant genotype effect}
}
\value{
A named list of ggplot objects.
}
\description{
Generate boxplots of the trait for each significant GWAS SNP
}
