% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snplinkage_subCodes.R
\name{combine_plots}
\alias{combine_plots}
\title{Function to combine multiple ggplot objects into a single gtable object}
\usage{
combine_plots(plots, labels_colname, title)
}
\arguments{
\item{plots}{List of ggplot objects}

\item{labels_colname}{Column name for SNP labels}

\item{title}{Title for the plot}
}
\value{
gtable
}
\description{
Function to combine multiple ggplot objects into a single gtable object
}
