% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLDcluster.R
\name{computeLDclusterByChr}
\alias{computeLDclusterByChr}
\title{Compute LD matrix for a cluster}
\usage{
computeLDclusterByChr(hapmap, haplotype_clusters, chr, cls, outfolder)
}
\arguments{
\item{hapmap}{a hapmap object containing the genotypic data}

\item{haplotype_clusters}{a list of haplotype clusters}

\item{chr}{the chromosome}

\item{cls}{the cluster}

\item{outfolder}{the output folder}
}
\description{
Compute LD matrix for a cluster
}
