% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_blocks.R
\name{computeLDclusters}
\alias{computeLDclusters}
\title{Compute LD matrices for the distance clusters}
\usage{
computeLDclusters(hapmap, haplotype_clusters)
}
\arguments{
\item{hapmap}{A list of HapMap data frames (from [readHapmap()])}

\item{haplotype_clusters}{A list of distance clusters (from [getDistClusters()])}
}
\value{
A list with the LD folder (`ld_folder`), cluster information
  (`out_info`) and the LD matrices (`matrices`).
}
\description{
Computes the r2 (squared correlation of allele dosages) between all SNPs of
each window and saves the matrices in the `LD_matrices` folder.
}
