% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.R
\name{convertGenoBi2Numeric}
\alias{convertGenoBi2Numeric}
\title{Convert HapMap genotypes to allele dosages}
\usage{
convertGenoBi2Numeric(habmap)
}
\arguments{
\item{habmap}{A HapMap data frame (11 annotation columns followed by
genotype columns), or a character matrix of genotype calls.}
}
\value{
An integer matrix (SNPs x accessions) counting copies of the minor
  allele (0, 1, 2); missing calls are `NA`.
}
\description{
Convert HapMap genotypes to allele dosages
}
\examples{
g <- data.frame(matrix(NA, 2, 11), G1 = c("AA", "CC"), G2 = c("AG", "CT"), G3 = c("GG", "NN"))
convertGenoBi2Numeric(g)
}
