% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHapCombBoxPlots.R
\name{create_dirs}
\alias{create_dirs}
\title{Create directories for haplotype combination boxplots}
\usage{
create_dirs(outfolder)
}
\arguments{
\item{outfolder}{The output folder where the directories will be created}
}
\value{
A list containing the paths of the created directories for groups a, b, and c
}
\description{
Create directories for haplotype combination boxplots
}
