% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_t_test.R
\name{evaluate_test_group0}
\alias{evaluate_test_group0}
\title{Evaluate haplotype combinations versus group 0 in each trait}
\usage{
evaluate_test_group0(t_test_snpComp)
}
\arguments{
\item{t_test_snpComp}{A list of tibbles containing t-test results for different SNP combinations}
}
\value{
A data frame with additional columns 'eval0' and 'dir0' indicating the evaluation and direction of group 0 comparison
}
\description{
Evaluate haplotype combinations versus group 0 in each trait
}
