% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.R
\name{extract_hapmap}
\alias{extract_hapmap}
\title{Extract HapMap data for significant SNPs}
\usage{
extract_hapmap(hapmap, gwas)
}
\arguments{
\item{hapmap}{A list of HapMap data frames (from [readHapmap()])}

\item{gwas}{A list of GWAS data frames (from [filter_gwas_data()])}
}
\value{
A list of HapMap data frames containing only the GWAS SNPs.
}
\description{
Extract HapMap data for significant SNPs
}
