% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{filter_gwas_data}
\alias{filter_gwas_data}
\title{Filter GWAS data by FDR threshold}
\usage{
filter_gwas_data(gwas)
}
\arguments{
\item{gwas}{Data frame with the GWAS data (from [read_gwas_file()])}
}
\value{
A list of data frames, one per chromosome, with the significant SNPs.
}
\description{
Filter GWAS data by FDR threshold
}
\examples{
gwas <- read_gwas_file(system.file("extdata", "gwas_area_ann19.csv.gz", package = "HaploTraitR"))
set_config(list(fdr_threshold = 0.1))
sig <- filter_gwas_data(gwas)
reset_config()
}
