% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.R
\name{find_nearby_snps}
\alias{find_nearby_snps}
\title{Find nearby SNPs to a given SNP list assuming they are in the same chromosome}
\usage{
find_nearby_snps(snp_list1, snp_list2, threshold)
}
\arguments{
\item{snp_list1}{A vector of SNP positions (reference)}

\item{snp_list2}{A vector of SNP positions (query)}

\item{threshold}{The maximum distance between the SNPs}
}
\value{
A list with, for each position in `snp_list1`, the positions of
  `snp_list2` within `threshold`.
}
\description{
Find nearby SNPs to a given SNP list assuming they are in the same chromosome
}
\examples{
find_nearby_snps(c(1000, 2000, 3000), c(1500, 2500, 3500, 5000), 1000)
}
