% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_haplotypes.R
\name{generateHapCombBoxPlots}
\alias{generateHapCombBoxPlots}
\title{Generate and save the haplotype trait plots of all blocks}
\usage{
generateHapCombBoxPlots(
  SNPcombTables,
  t_test_snpComp,
  pwidth = NULL,
  pheight = NULL
)
}
\arguments{
\item{SNPcombTables}{The accession table (from [getSNPcombTables()])}

\item{t_test_snpComp}{Pairwise tests (from [testSNPcombs()])}

\item{pwidth, pheight}{Not used anymore; the size adapts to the number of haplotypes.}
}
\value{
Invisibly, a named list of ggplot objects.
}
\description{
Plots are saved in `plots/haplotype_trait_plots/significant` or
`.../not_significant` according to the overall test of the block.
}
