% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config}
\alias{get_config}
\title{Get a configuration parameter}
\usage{
get_config(name)
}
\arguments{
\item{name}{The name of the configuration parameter}
}
\value{
The value of the configuration parameter
}
\description{
Get a configuration parameter
}
\examples{
get_config("dist_threshold")
}
