% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.R
\name{get_pheno_geno}
\alias{get_pheno_geno}
\title{Get the phenotype and genotype data}
\usage{
get_pheno_geno(subhapmap, pheno)
}
\arguments{
\item{subhapmap}{A list of HapMap data frames (from [extract_hapmap()])}

\item{pheno}{A phenotype data frame (accession, trait)}
}
\value{
A long data frame with columns `Phenotype`, `variable` (SNP),
  `value` (genotype call) and `Sample`.
}
\description{
Get the phenotype and genotype data
}
