% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_blocks.R
\name{getDistClusters}
\alias{getDistClusters}
\title{Get distance clusters around the significant GWAS SNPs}
\usage{
getDistClusters(gwas, hapmap)
}
\arguments{
\item{gwas}{A list of GWAS data frames split by chromosome (from [filter_gwas_data()])}

\item{hapmap}{A list of HapMap data frames split by chromosome (from [readHapmap()])}
}
\value{
A list (per chromosome) of named lists (per GWAS SNP) of SNP positions.
}
\description{
For every significant GWAS SNP, collects the genotyped SNPs within
`dist_threshold` bp. Windows with fewer than `dist_cluster_count` SNPs are dropped.
}
