% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHaplotypeClusters.R
\name{getHaplotypeClustersByChr}
\alias{getHaplotypeClustersByChr}
\title{Get haplotype clusters from GWAS and hapmap data By chromosome}
\usage{
getHaplotypeClustersByChr(
  gwas,
  hapmap,
  chromosome,
  dist_threshold,
  dist_cluster_count
)
}
\arguments{
\item{gwas}{A list of GWAS data frames split by chromosome}

\item{hapmap}{A list of hapmap data frames split by chromosome}

\item{chromosome}{The chromosome of interest}

\item{dist_threshold}{The maximum distance between SNPs to consider for clustering}

\item{dist_cluster_count}{The minimum number of SNPs required to form a cluster}
}
\value{
A list of haplotype clusters
}
\description{
Get haplotype clusters from GWAS and hapmap data By chromosome
}
