% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_blocks.R
\name{getLDclusters}
\alias{getLDclusters}
\title{Cluster SNPs based on LD matrices}
\usage{
getLDclusters(LDsInfo)
}
\arguments{
\item{LDsInfo}{A list returned by [computeLDclusters()] or [retrieveLDMatricesFromFolder()]}
}
\value{
A named list (per GWAS SNP) of the SNP identifiers in its LD block.
}
\description{
SNPs are linked when their r2 is at least `ld_threshold`; the connected
group of SNPs that contains the GWAS SNP forms its LD block. Blocks smaller
than `dist_cluster_count` SNPs are discarded.
}
