% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaploTraitR-package.R
\docType{package}
\name{HaploTraitR-package}
\alias{HaploTraitR}
\alias{HaploTraitR-package}
\title{HaploTraitR: Haplotype-Based Trait Analysis for Plant Breeding}
\description{
HaploTraitR turns GWAS hits into haplotype blocks (SNPs linked by LD),
tests how each haplotype affects a trait, and reports the superior
haplotypes, diagnostic markers and candidate parents in plots and a
formatted Excel workbook.
}
\details{
The quickest way to use the package is [run_haplotraitr()]. The individual
steps ([read_gwas_file()], [readHapmap()], [getDistClusters()], ...) remain
available for custom workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlsammanAlsamman/HaploTraitR}
  \item \url{https://alsammanalsamman.github.io/HaploTraitR/}
  \item Report bugs at \url{https://github.com/AlsammanAlsamman/HaploTraitR/issues}
}

}
\author{
\strong{Maintainer}: Alsamman Alsamman \email{aalsamman100@gmail.com}

Other contributors:
\itemize{
  \item Tamara Ortiz \email{tamara.ortizru@gmail.com} [contributor]
  \item Khaled Helmy \email{khaledhelmy444@gmail.com} [contributor]
  \item Zakaria Kehel \email{z.kehel@cgiar.org} [contributor]
  \item ICARDA [copyright holder]
}

}
\keyword{internal}
