% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_t_test.R
\name{merge_with_list_name}
\alias{merge_with_list_name}
\title{Merge a list of data frames with a list name}
\usage{
merge_with_list_name(df_list, pheno)
}
\arguments{
\item{df_list}{A list of data frames to merge}

\item{pheno}{The phenotype name to add to the merged data frame}
}
\value{
A merged data frame with an added phenotype column
}
\description{
Merge a list of data frames with a list name
}
