% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snplinkage_subCodes.R
\name{plot_snp_positions}
\alias{plot_snp_positions}
\title{Function to plot SNP positions}
\usage{
plot_snp_positions(
  df_snp,
  upper_subset = NULL,
  labels_colname = NULL,
  colors = snp_position_colors(nrow(df_snp))
)
}
\arguments{
\item{df_snp}{Data frame with columns position, labels_colname}

\item{upper_subset}{Subset of SNPs to highlight}

\item{labels_colname}{Column name for SNP labels}

\item{colors}{Colors for SNP positions}
}
\value{
ggplot
}
\description{
Function to plot SNP positions
}
