% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blocks.R
\name{plotCombMatrix}
\alias{plotCombMatrix}
\title{Plot the haplotype allele matrix of a block}
\usage{
plotCombMatrix(
  cluster_id,
  haplotypes,
  gwas = NULL,
  snps = NULL,
  outfolder = NULL
)
}
\arguments{
\item{cluster_id}{A block id or GWAS SNP}

\item{haplotypes}{A haplotype table (from [convertLDclusters2Haps()] or [getHapCombSamples()])}

\item{gwas}{Not used; kept for compatibility}

\item{snps}{Not used; kept for compatibility}

\item{outfolder}{Folder in which to save the plot (`NULL`: do not save)}
}
\value{
A ggplot object.
}
\description{
Plot the haplotype allele matrix of a block
}
