% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHapCombDistribution.R
\name{plotHapCombDistribution}
\alias{plotHapCombDistribution}
\title{Plotting Haplotype Combination Distribution for specific SNP}
\usage{
plotHapCombDistribution(snp, SNPcombTables)
}
\arguments{
\item{snp}{The SNP of interest}

\item{SNPcombTables}{A list of data frames containing haplotype combinations and phenotype values}
}
\description{
Plotting Haplotype Combination Distribution for specific SNP
}
