% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blocks.R
\name{plotLDForClusters}
\alias{plotLDForClusters}
\title{Plot LD heat maps for all LD blocks}
\usage{
plotLDForClusters(clusterLDs, LDsInfo = NULL)
}
\arguments{
\item{clusterLDs}{A list of LD blocks (from [getLDclusters()])}

\item{LDsInfo}{Optional output of [computeLDclusters()]; otherwise the
matrices are read from the `LD_matrices` folder of the output folder}
}
\value{
Invisibly, a list of ggplot objects (also saved in `plots/LD_heatmaps`).
}
\description{
Plot LD heat maps for all LD blocks
}
