% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blocks.R
\name{plotLDheatmap}
\alias{plotLDheatmap}
\title{Plot an LD heat map}
\usage{
plotLDheatmap(LD_matrix)
}
\arguments{
\item{LD_matrix}{A square LD (r2) matrix whose row names are `chr:pos`}
}
\value{
A ggplot object.
}
\description{
Plot an LD heat map
}
