% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGWAS.R
\name{readGWAS}
\alias{readGWAS}
\title{read gwas results file rs, chr, pos, p}
\usage{
readGWAS(gwasfile, header = TRUE, sep = "\\t")
}
\arguments{
\item{gwasfile}{Path to the GWAS file}
}
\value{
A data frame with the GWAS data
}
\description{
read gwas results file rs, chr, pos, p
}
\examples{
gwasfile<-"SignificantSNP_GWAS.csv"
gwas<-readGWAS(gwasfile)
print(gwas)
}
