% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{reset_config}
\alias{reset_config}
\title{Reset the configuration to default}
\usage{
reset_config()
}
\value{
Invisibly `NULL`.
}
\description{
Reset the configuration to default
}
\examples{
reset_config()
}
