% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_blocks.R
\name{retrieveLDMatricesFromFolder}
\alias{retrieveLDMatricesFromFolder}
\title{Get LD matrix information from a folder}
\usage{
retrieveLDMatricesFromFolder(folder)
}
\arguments{
\item{folder}{The folder where the LD matrices are stored (written by [computeLDclusters()])}
}
\value{
A list as returned by [computeLDclusters()].
}
\description{
Get LD matrix information from a folder
}
