% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLDs2folder.R
\name{saveLDs2folder}
\alias{saveLDs2folder}
\title{save LD matrices to a folder}
\usage{
saveLDs2folder(LDsInfo, outfolder)
}
\arguments{
\item{LDsInfo}{a list containing the LD matrices and the info}

\item{outfolder}{the folder to save the LD matrices}
}
\description{
save LD matrices to a folder
}
