% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{saveTTestResultsToFile}
\alias{saveTTestResultsToFile}
\title{Save the haplotype test results to a file}
\usage{
saveTTestResultsToFile(t_test_snpComp)
}
\arguments{
\item{t_test_snpComp}{The list returned by [testSNPcombs()]}
}
\value{
Invisibly, the path of the CSV file.
}
\description{
Save the haplotype test results to a file
}
