% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_config}
\alias{set_config}
\title{Set configuration parameters for HaploTraitR}
\usage{
set_config(config = list())
}
\arguments{
\item{config}{A named list of configuration parameters. See [list_config()]
for the available parameters.}
}
\value{
Invisibly, the previous values of the changed parameters.
}
\description{
Set configuration parameters for HaploTraitR
}
\examples{
set_config(list(dist_threshold = 2000000, dist_cluster_count = 10))
reset_config()
}
