% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_geno_pheno.R
\name{boxplot_genotype_phenotype}
\alias{boxplot_genotype_phenotype}
\title{Generate boxplots for genotype-phenotype data}
\usage{
boxplot_genotype_phenotype(
  genotype_phenotype_data,
  y_label = "Phenotype Value",
  outfolder = NULL,
  method = "t.test"
)
}
\arguments{
\item{genotype_phenotype_data}{A data frame containing genotype-phenotype data generated by get_pheno_geno()}

\item{y_label}{The label for the y-axis}

\item{outfolder}{An optional output folder for saving the plots}

\item{method}{The method for comparing means, can be "t.test" or "wilcox.test"}
}
\value{
A list of ggplot2 objects or NULL if outfolder is specified
}
\description{
Generate boxplots for genotype-phenotype data
}
