% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_geno_pheno.R
\name{boxplot_genotype_phenotype}
\alias{boxplot_genotype_phenotype}
\title{Generate boxplots for genotype-phenotype data}
\usage{
boxplot_genotype_phenotype(
  pheno,
  gwas,
  hapmap,
  plot_width = 5,
  plot_height = 5
)
}
\arguments{
\item{pheno}{A data frame containing phenotype data}

\item{gwas}{A data frame containing GWAS data}

\item{hapmap}{A data frame containing hapmap data}
}
\value{
A list of ggplot2 objects or NULL if outfolder is specified
}
\description{
Generate boxplots for genotype-phenotype data
}
