% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLDcluster.R
\name{computeLDclusters}
\alias{computeLDclusters}
\title{Compute LD matrix for clusters}
\usage{
computeLDclusters(hapmap, haplotype_clusters)
}
\arguments{
\item{hapmap}{a hapmap object containing the genotypic data}

\item{haplotype_clusters}{a list of haplotype clusters}
}
\value{
a list containing the path to the folder with LD matrices and information about the clusters
}
\description{
Compute LD matrix for clusters
}
