% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLDcluster.R
\name{computeLDclusters}
\alias{computeLDclusters}
\title{Compute LD matrix for a clusters}
\usage{
computeLDclusters(hapmap, haplotype_clusters, outfolder)
}
\arguments{
\item{hapmap}{a hapmap object containing the genotypic data}

\item{haplotype_clusters}{a list of haplotype clusters}

\item{outfolder}{the output folder}
}
\value{
a folder containing the LD matrices
}
\description{
Compute LD matrix for a clusters
}
