% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDclusters2Haps.R
\name{convertLDclusters2Haps}
\alias{convertLDclusters2Haps}
\title{Convert LD clusters to haplotypes and calculate the frequency of each haplotype combination}
\usage{
convertLDclusters2Haps(hapmap, clusterLDs, savecopy = TRUE)
}
\arguments{
\item{hapmap}{a hapmap object containing the genotypic data}

\item{clusterLDs}{a list of clusters}

\item{savecopy}{a boolean indicating whether to save a copy of the haplotype combinations to the project folder}
}
\value{
a data frame containing the haplotype combinations and their frequencies
}
\description{
Convert LD clusters to haplotypes and calculate the frequency of each haplotype combination
}
