% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDclusters2Haps.R
\name{convertLDclusters2Haps}
\alias{convertLDclusters2Haps}
\title{Convert LD clusters to haplotypes and calculate the frequency of each haplotype combination}
\usage{
convertLDclusters2Haps(hapmap, clusterLDs, comb_freq_threshold)
}
\arguments{
\item{hapmap}{A hapmap object}

\item{clusterLDs}{A list of LD clusters}

\item{comb_freq_threshold}{The minimum frequency of a haplotype combination}
}
\value{
A list of haplotype combinations with their frequencies
}
\description{
Convert LD clusters to haplotypes and calculate the frequency of each haplotype combination
}
