% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_unique_result_folder.R
\name{create_unique_result_folder}
\alias{create_unique_result_folder}
\title{Create a unique result folder for HaplotraitR}
\usage{
create_unique_result_folder(
  base_folder_name = "haplotraitR_run",
  location = NULL
)
}
\arguments{
\item{base_folder_name}{The base name for the result folder}

\item{location}{The location where the folder should be created}
}
\value{
The path to the created folder
}
\description{
Create a unique result folder for HaplotraitR
}
\examples{
result_folder <- create_unique_result_folder()
cat("Results will be stored in:", result_folder, "\n")
}
