% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hapmap.R
\name{extract_hapmap}
\alias{extract_hapmap}
\title{Extract hapmap data for significant SNPs}
\usage{
extract_hapmap(hapmap, gwas)
}
\arguments{
\item{hapmap}{A list of hapmap data frames}

\item{gwas}{A list of GWAS data frames}
}
\value{
A list of hapmap data frames for significant SNPs
}
\description{
Extract hapmap data for significant SNPs
}
