% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGWAS.R
\name{filter_gwas_data}
\alias{filter_gwas_data}
\title{Filter GWAS data by FDR threshold}
\usage{
filter_gwas_data(gwas)
}
\arguments{
\item{gwas}{Data frame with the GWAS data}
}
\value{
A list of data frames split by chromosome with filtered GWAS data
}
\description{
Filter GWAS data by FDR threshold
}
\examples{
gwasfile <- "SignificantSNP_GWAS.csv"
gwas <- read_gwas_file(gwasfile)
filtered_gwas <- filter_gwas_data(gwas)
print(filtered_gwas)
}
