% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearbySNPs.R
\name{find_nearby_snps}
\alias{find_nearby_snps}
\title{Find nearby SNPs to a given SNP list assuming they are in the same chromosome}
\usage{
find_nearby_snps(snp_list1, snp_list2, threshold)
}
\arguments{
\item{snp_list1}{A list of SNPs (Reference)}

\item{snp_list2}{A list of SNPs (Query)}

\item{threshold}{The maximum distance between the SNPs}
}
\value{
A list of SNPs from snp_list2 that are within the threshold distance from snp_list1
}
\description{
Find nearby SNPs to a given SNP list assuming they are in the same chromosome
}
\examples{
list1 <- c(1000, 2000, 3000)
list2 <- c(1500, 2500, 3500, 5000)
threshold <- 1000
result <- find_nearby_snps(list1, list2, threshold)
print(result)
}
