% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHapCombBoxPlots.R
\name{generateHapCombBoxPlots}
\alias{generateHapCombBoxPlots}
\title{Generate and save haplotype combination boxplots}
\usage{
generateHapCombBoxPlots(SNPcombTables, t_test_snpComp, pwidth = 6, pheight = 6)
}
\arguments{
\item{SNPcombTables}{A list containing the SNP combination tables}

\item{t_test_snpComp}{A list containing the t-test results for SNP comparisons}

\item{pwidth}{The width of the saved plots (default is 6)}

\item{pheight}{The height of the saved plots (default is 6)}
}
\description{
Generate and save haplotype combination boxplots
}
