% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pheno_geno.R
\name{get_pheno_geno}
\alias{get_pheno_geno}
\title{Get the phenotype and genotype data}
\usage{
get_pheno_geno(subhapmap, pheno)
}
\arguments{
\item{subhapmap}{A subset of the hapmap data containing the significant SNPs}

\item{pheno}{A phenotype data frame}
}
\value{
A data frame with the phenotype and genotype data
}
\description{
Get the phenotype and genotype data
}
