% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHaplotypeClusters.R
\name{getDistClusters}
\alias{getDistClusters}
\title{Get haplotype clusters from GWAS and hapmap data}
\usage{
getDistClusters(gwas, hapmap)
}
\arguments{
\item{gwas}{A list of GWAS data frames split by chromosome}

\item{hapmap}{A list of hapmap data frames split by chromosome}
}
\value{
A list of haplotype clusters
}
\description{
Get haplotype clusters from GWAS and hapmap data
}
