% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHaplotypeClusters.R
\name{getDistClustersByChr}
\alias{getDistClustersByChr}
\title{Get Distance clusters from GWAS and hapmap data By chromosome}
\usage{
getDistClustersByChr(gwas, hapmap, chromosome)
}
\arguments{
\item{gwas}{A list of GWAS data frames split by chromosome}

\item{hapmap}{A list of hapmap data frames split by chromosome}

\item{chromosome}{The chromosome of interest}
}
\value{
A list of Dist clusters
}
\description{
Get Distance clusters from GWAS and hapmap data By chromosome
}
