% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHapCombSamples.R
\name{getHapCombSamples}
\alias{getHapCombSamples}
\title{Get samples for a haplotype combination}
\usage{
getHapCombSamples(haplotypes, hapmap)
}
\arguments{
\item{haplotypes}{A list of haplotype combinations}

\item{hapmap}{A hapmap object}
}
\value{
A list of samples for each haplotype combination
}
\description{
Get samples for a haplotype combination
}
