% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterLD.R
\name{getLDclusters}
\alias{getLDclusters}
\title{Cluster SNPs based on LD matrices}
\usage{
getLDclusters(LDsInfo)
}
\arguments{
\item{LDsInfo}{a list containing the LD matrices and the info of the clusters}
}
\value{
LD cluster information
}
\description{
Cluster SNPs based on LD matrices
}
