% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterLD.R
\name{getLDclusters}
\alias{getLDclusters}
\title{Cluster SNPs based on LD matrices}
\usage{
getLDclusters(LDsInfo, ld_threshold, cls_count = 3)
}
\arguments{
\item{LDsInfo}{a list containing the LD matrices and the info of the clusters}

\item{ld_threshold}{The minimum LD value to consider two SNPs in LD}
}
\value{
LD cluster information
}
\description{
Cluster SNPs based on LD matrices
}
