% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSNPcombTables.R
\name{getSNPcombTables}
\alias{getSNPcombTables}
\title{Get SNP combinations sample tables}
\usage{
getSNPcombTables(haplotypes, pheno, savecopy = TRUE)
}
\arguments{
\item{haplotypes}{A list of haplotype combinations}

\item{pheno}{A phenotype data frame}

\item{savecopy}{A boolean indicating whether to save a copy of the sample tables to the project folder+}
}
\value{
A list of sample tables for each SNP combination
}
\description{
Get SNP combinations sample tables
}
