% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSNPcombTables.R
\name{getSNPcombTables}
\alias{getSNPcombTables}
\title{Get SNP combinations sample tables}
\usage{
getSNPcombTables(haplotypes, hapmap)
}
\arguments{
\item{haplotypes}{A list of haplotype combinations}

\item{hapmap}{A hapmap object}
}
\value{
A list of sample tables for each SNP combination
}
\description{
Get SNP combinations sample tables
}
