% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_config}
\alias{load_config}
\title{Load configuration from a file}
\usage{
load_config(filename)
}
\arguments{
\item{filename}{The name of the file to load the configuration from}
}
\description{
Load configuration from a file
}
\examples{
load_config("config.rds")
load_config("config.txt")
}
