% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_plot.R
\name{plotCombMatrix}
\alias{plotCombMatrix}
\title{Plot Haplotype Combinations Matrix}
\usage{
plotCombMatrix(cluster_id, haplotypes, gwas, snps = NULL, outfolder = NULL)
}
\arguments{
\item{cluster_id}{cluster id}

\item{haplotypes}{a data frame containing haplotype information}

\item{gwas}{a list of GWAS SNPs}

\item{snps}{a vector of SNP names to plot}

\item{outfolder}{a folder to save the plots}
}
\value{
a ggplot objects showing the haplotype combinations matrix
}
\description{
Plot Haplotype Combinations Matrix
}
