% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHapCombBoxPlot.R
\name{plotHapCombBoxPlot}
\alias{plotHapCombBoxPlot}
\title{Plotting Haplotype Combination Boxplot for specific SNP}
\usage{
plotHapCombBoxPlot(
  cls_snp = NULL,
  SNPcombTables,
  t_test_snpComp,
  with_dots = TRUE,
  outfolder = ""
)
}
\arguments{
\item{cls_snp}{The SNP of interest, if NULL then plots for all SNPs in t_test_snpComp are generated}

\item{SNPcombTables}{A list of data frames containing haplotype combinations and phenotype values}

\item{t_test_snpComp}{A list of t-test results for haplotype combinations}
}
\description{
Plotting Haplotype Combination Boxplot for specific SNP
}
