% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHapCombBoxPlot.R
\name{plotHapCombBoxPlot}
\alias{plotHapCombBoxPlot}
\title{Plotting Haplotype Combination Boxplot for specific SNP}
\usage{
plotHapCombBoxPlot(snp, SNPcombTables, t_test_snpComp)
}
\arguments{
\item{snp}{The SNP of interest}

\item{SNPcombTables}{A list of data frames containing haplotype combinations and phenotype values}

\item{t_test_snpComp}{A list of t-test results for haplotype combinations}
}
\description{
Plotting Haplotype Combination Boxplot for specific SNP
}
