% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHapCombBoxPlot.R
\name{plotHapCombBoxPlot}
\alias{plotHapCombBoxPlot}
\title{Plotting Haplotype Combination Boxplot for specific SNP}
\usage{
plotHapCombBoxPlot(cls_snp, SNPcombTables, t_test_snpComp, with_dots = TRUE)
}
\arguments{
\item{SNPcombTables}{A list of data frames containing haplotype combinations and phenotype values}

\item{t_test_snpComp}{A list of t-test results for haplotype combinations}

\item{snp}{The SNP of interest}
}
\description{
Plotting Haplotype Combination Boxplot for specific SNP
}
