% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_plot.R
\name{plotLDCombMatrix}
\alias{plotLDCombMatrix}
\title{Plot LD and Haplotype with Combination Matrix}
\usage{
plotLDCombMatrix(clusterLDs, haplotypes, gwas)
}
\arguments{
\item{clusterLDs}{a list of LD clusters}

\item{haplotypes}{a data frame containing haplotype information}

\item{gwas}{a list of GWAS SNPs}

\item{ld_matrix_folder}{a folder containing LD matrices}

\item{snps}{a vector of SNP names to plot}

\item{outfolder}{a folder to save the plots}
}
\description{
Plot LD and Haplotype with Combination Matrix
}
