% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_plot.R
\name{plotLDCombMatrix}
\alias{plotLDCombMatrix}
\title{Plot LD and Haplotype with Combination Matrix}
\usage{
plotLDCombMatrix(
  ld_matrix_folder,
  clusterLDs,
  haplotypes,
  gwas,
  cluster_ids = NULL,
  outfolder = NULL
)
}
\arguments{
\item{ld_matrix_folder}{a folder containing LD matrices}

\item{clusterLDs}{a list of LD clusters}

\item{haplotypes}{a data frame containing haplotype information}

\item{gwas}{a list of GWAS SNPs}

\item{outfolder}{a folder to save the plots}

\item{snps}{a vector of SNP names to plot}
}
\description{
Plot LD and Haplotype with Combination Matrix
}
