% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_plot.R
\name{plotLDheatmap}
\alias{plotLDheatmap}
\title{Plotting LD Heatmap for a Specific Haplotype}
\usage{
plotLDheatmap(LD_matrix)
}
\arguments{
\item{LD_matrix}{The LD matrix.}
}
\value{
A ggplot object showing the LD heatmap.
}
\description{
This function plots an LD heatmap for a specific haplotype using an LD matrix and hapmap object.
}
