% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGWAS.R
\name{read_gwas_file}
\alias{read_gwas_file}
\title{Read GWAS results file with specified columns}
\usage{
read_gwas_file(gwasfile, header = TRUE, sep = "\\t")
}
\arguments{
\item{gwasfile}{Path to the GWAS file}

\item{header}{Logical, whether the file contains a header row}

\item{sep}{Separator used in the file}
}
\value{
A data frame with the GWAS data
}
\description{
Read GWAS results file with specified columns
}
\examples{
gwasfile <- "SignificantSNP_GWAS.csv"
gwas <- read_gwas_file(gwasfile)
print(gwas)
}
