% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLDcluster.R
\name{retrieveLDMatricesFromFolder}
\alias{retrieveLDMatricesFromFolder}
\title{Get LD matrix information from a folder}
\usage{
retrieveLDMatricesFromFolder(folder)
}
\arguments{
\item{folder}{The folder where the LD matrices are stored}
}
\value{
A list containing the path to the folder with LD matrices and information about the clusters
}
\description{
Get LD matrix information from a folder
}
