% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{save_config}
\alias{save_config}
\title{Save configuration to a file}
\usage{
save_config(filename)
}
\arguments{
\item{filename}{The name of the file to save the configuration to}
}
\description{
Save configuration to a file
}
\examples{
save_config("config.txt")
}
