% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_t_test.R
\name{saveTTestResultsToFile}
\alias{saveTTestResultsToFile}
\title{Convert t-test results to a data frame and save to a file}
\usage{
saveTTestResultsToFile(t_test_snpComp)
}
\arguments{
\item{t_test_snpComp}{A list containing the t-test results for SNP comparisons}

\item{outfolder}{The output folder for saving the file}
}
\description{
Convert t-test results to a data frame and save to a file
}
