% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_config}
\alias{set_config}
\title{Set configuration parameters for HaploTraitR}
\usage{
set_config(config = list())
}
\arguments{
\item{config}{A named list of configuration parameters.}
}
\description{
Set configuration parameters for HaploTraitR
}
\examples{
config <- list(dist_threshold = 2000000, dist_cluster_count = 10)
set_config(config)
}
